refactor: tech trainers feedback
refactored codebase based on tech trainer feedback
