+++ Gemfile
@@ -1,3 +1,3 @@
-ruby '2.2.3'
+ruby '2.3.1'
 source 'https://rubygems.org'
 gemspec
