+++ lib/micro_mvc_ruby/orm/database.rb
@@ -1,0 +1,13 @@
+module MicroMvcRuby
+  class Database
+    attr_accessor :db
+
+    def initialize(db_file = 'app.db') 
+      @db = SQLite3::Database.new File.join db_file 
+    end 
+
+    def self.execute(query, *args)
+      new.db.execute(query, args)
+    end
+  end
+end
