+++ lib/micro_mvc_ruby/utils/class_patching.rb
@@ -1,0 +1,9 @@
+class Class 
+  def getter_setter(*args) 
+    args.each do |arg|
+     self.class_eval("def #{arg}; @#{arg};end") 
+     self.class_eval("def #{arg}=(val); @#{arg}=val; end") 
+    end
+  end
+end
+
